requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
orjson==3.8.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
sqlparse==0.4.3
starlette==0.20.4
typing_extensions==4.4.0
tzdata==2022.5
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.1
websockets==10.4
